refactor(cli): migrate facade imports to lib.api
All eight CLI command modules now route their domain-, ssh-, and
auth-related imports through ``terok.lib.api`` instead of the
``terok.lib.domain.facade`` shim.  CLI and TUI now share the same single
boundary.

Mechanical replacement: ``from ...lib.domain.facade import X`` becomes
``from ...lib.api import X``.  No symbol-name changes; api.py exports
the same surface facade did.

tach.toml: terok.cli no longer depends on ``terok.lib.domain.facade``;
``terok.lib.api`` takes its place.
