+++ src/terok/cli/commands/task.py
@@ -12,8 +12,7 @@ from typing import Any
 
 from terok_executor import PROVIDER_NAMES as _PROVIDER_NAMES
 
-from ...lib.core.config import get_logs_partial_streaming as _get_logs_partial_streaming
-from ...lib.domain.facade import (
+from ...lib.api import (
     HeadlessRunRequest,
     LogViewOptions,
     build_images,
@@ -34,6 +33,7 @@ from ...lib.domain.facade import (
     task_status,
     task_stop,
 )
+from ...lib.core.config import get_logs_partial_streaming as _get_logs_partial_streaming
 from ...lib.orchestration.tasks import resolve_task_id
 from ._completers import add_project_id, add_task_id, complete_preset_names, set_completer
