+++ tach.toml
@@ -46,6 +46,7 @@ exclude = ["pydevd_pycharm", "pydevd"]
 path = "terok.cli"
 layer = "presentation"
 depends_on = [
+    "terok.lib.api",
     "terok.lib.orchestration.agent_config",
     "terok.lib.domain.panic",
     "terok.lib.orchestration.container_doctor",
@@ -56,7 +57,6 @@ depends_on = [
     "terok.lib.core.runtime",
     "terok.lib.core.version",
     "terok.lib.core.yaml_schema",
-    "terok.lib.domain.facade",
     "terok.lib.domain.storage",
     "terok.lib.util.ansi",
     "terok.lib.util.emoji",
